# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation de l'API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/ressources": {
            "get": {
                "tags": [
                    "Ressources"
                ],
                "summary": "Liste des ressources",
                "description": "Récupère la liste de toutes les ressources.",
                "operationId": "3e11064f9d5e9e717e3574b719cb52cd",
                "responses": {
                    "200": {
                        "description": "Liste des ressources",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "schema": "RessourcesListResponse"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Aucune ressource disponible",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Message d'erreur",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Ressources"
                ],
                "summary": "Création d'une ressource",
                "description": "Crée une nouvelle ressource.",
                "operationId": "ca8262df979c518aa68dca59e19a4866",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "schema": "RessourceCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ressource créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Message de confirmation",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Mauvaise requête",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Message d'erreur",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Échec de la validation des données",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "description": "Liste des erreurs de validation",
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/ressources/{id}": {
            "get": {
                "tags": [
                    "Ressources"
                ],
                "summary": "Affichage d'une ressource",
                "description": "Affiche les détails d'une ressource spécifiée par son ID.",
                "operationId": "1de62c82ef1dbab15585ae77a3582b12",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la ressource",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de la ressource",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "schema": "RessourceShowRequest"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ressource non trouvée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Message d'erreur",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Ressources"
                ],
                "summary": "Mise à jour d'une ressource",
                "description": "Met à jour les informations d'une ressource spécifiée par son ID.",
                "operationId": "c29ab2056e1cab506e291dde1b1db123",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la ressource",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Données de la ressource à mettre à jour",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "schema": "RessourceUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ressource mise à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Message de succès",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ressource non trouvée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Message d'erreur",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Ressources"
                ],
                "summary": "Suppression d'une ressource",
                "description": "Supprime une ressource spécifiée par son ID.",
                "operationId": "4a1fc2946fccf29d9338f5bae092665e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la ressource",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ressource supprimée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Message de confirmation",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ressource non trouvée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Message d'erreur",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}